稍微修改所有的电力参数数据计算 by PgNat

+++ src/main/java/power/api/controller/responseModel/powerMonitoring/PhaseCurrentResponse.java
@@ -7,9 +7,9 @@ import java.sql.Timestamp;
 @Data
 public class PhaseCurrentResponse {
     private Timestamp createAt;
-    private double phaseCurrentA;
-    private double phaseCurrentB;
-    private double phaseCurrentC;
+    private float phaseCurrentA;
+    private float phaseCurrentB;
+    private float phaseCurrentC;
 
     public PhaseCurrentResponse() {
     }
